Refactor: Mv Network object to src
