+++ model/vanilla_deepQN/model.py
@@ -3,15 +3,17 @@ Vanilla DQN
 """
 import os
 import sys
+
+
 WORK_DIR = os.environ['ROOT_DIR']
 sys.path.append(WORK_DIR)
 
 import json
 import numpy as np
 from src.base_model.base_model import BaseModel
+from src.base_networks.base_network import Network
 import torch
 nn = torch.nn
-from collections import OrderedDict
 
 
 class Model(BaseModel):
@@ -104,35 +106,3 @@ class Model(BaseModel):
         return status
 
 
-class Network(nn.Module):
-    def __init__(self, nb_features, nb_actions, params, seed):
-        super(Network, self).__init__()
-        self.seed = torch.manual_seed(seed)
-        print('Cuda is available: {}'.format(torch.cuda.is_available()))
-        self.device = torch.device(
-            "cuda:0" if torch.cuda.is_available() else "cpu")
-
-        self.network = torch.nn.Sequential(
-            OrderedDict([
-                ('fc1', nn.Linear(in_features=nb_features,
-                                  out_features=params['network']['fc1'],
-                                  bias=True)),
-                ('relu1', nn.ReLU()),
-                ('fc2', nn.Linear(in_features=params['network']['fc1'],
-                                  out_features=params['network']['fc2'],
-                                  bias=True)),
-                ('relu2', nn.ReLU()),
-                ('fc3', nn.Linear(in_features=params['network']['fc2'],
-                                  out_features=nb_actions,
-                                  bias=True))]))
-
-    def forward(self, state):
-        """
-
-        Args:
-            state: torch.nn.Tensor object [batch_size, feature_size]
-
-        Returns:
-            target q_values for each action available
-        """
-        return self.network.forward(state)
